requirements.txt
alembic>=1.4.2
aniso8601>=8.0.0
beautifulsoup4>=4.9.1
certifi>=2020.6.20
chardet>=3.0.4
Click>=7.1.2
Flask>=1.1.2
Flask-JWT-Extended>=3.24.1
Flask-Login>=0.5.0
Flask-Migrate>=2.5.3
Flask-RESTful>=0.3.8
Flask-Script>=2.0.6
Flask-Sijax>=0.4.1
Flask-SQLAlchemy>=2.4.4
future>=0.18.2
idna>=2.10
itsdangerous>=2.0.0a1
Jinja2>=3.0.0a1
jsonify>=0.5
Mako>=1.1.3
MarkupSafe>=2.0.0a1
passlib>=1.7.2
psycopg2-binary>=2.8.5
PyJWT>=1.7.1
pyTelegramBotAPI>=3.7.2
python-dateutil>=2.8.1
python-dotenv>=0.14.0
python-editor>=1.0.4
pytz>=2020.1
requests==2.7.0
Sijax>=0.3.2
six>=1.15.0
soupsieve>=2.0.1
SQLAlchemy>=1.3.18
telebot>=0.0.3
telegram>=0.0.1
urllib3>=1.25.10
uwsgidecorators>=1.1.0
Werkzeug>=1.0.1
youtube-dl>=2020.6.16.1
setuptools>=49.2.0
selenium>=3.141.0